+++ Source/Motor/StepperUtility.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include <cstddef>
+#include <cstdint>
 
 namespace motor
 {
@@ -10,8 +11,8 @@ class StepperUtility
 public:
   StepperUtility(const size_t interruptRateHz, const size_t stepsPerRotation);
   size_t GetNextStepTick(const float stepsPerSecond, const size_t lastStepTick);
-  float DrpmToStepsPerSecond(const float drpm) const;
-  float StepsPerSecondToDrpm(const float steps) const;
+  float DrpmToStepsPerSecond(const int32_t drpm) const;
+  int32_t StepsPerSecondToDrpm(const float steps) const;
 
 private:
   const size_t InterruptRateHz;
